whatpm/t/webidl/webidl-typedef.dat

#data
typedef a DOMString;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef a DOMString;

#data
module m {
  typedef a DOMString;
};
#errors
2;typedef ignored;i
#document
| module m {
| 
| typedef a DOMString;
| 
| };

#data
typedef a _DOMString;
#errors
1;non-module definition;s

#data
typedef DOMString a;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef DOMString a;

#data
typedef _DOMString a;
#errors
1;non-module definition;s

#data
typedef m::DOMString a;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef m::DOMString a;

#data
typedef ::DOMString a;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef ::DOMString a;

#data
typedef ::m::DOMString a;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef ::m::DOMString a;

#data
typedef m::_DOMString a;
#errors
1;non-module definition;s

#data
typedef ::m::_DOMString a;
#errors
1;non-module definition;s

#data
typedef ::_DOMString a;
#errors
1;non-module definition;s

#data
typedef dom::DOMString DOMString;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef dom::DOMString DOMString;

#data
typedef sequence<unsigned  short> DOMString;
#errors
1;non-module definition;s
1;typedef ignored;i
#document
| typedef sequence<unsigned short> DOMString;

#data
module m {
  typedef dom::DOMString a;
  typedef long a;
};
#errors
3;duplicate qname;m;::m::a
2;typedef ignored;i

#data
module m {
  typedef long a;
  typedef dom::DOMString a;
};
#errors
3;duplicate qname;m;::m::a
3;typedef ignored;i

#data
module m {
  typedef dom::DOMString a;
  interface b {
    attribute a c;
  };
};
#errors
2;typedef ignored;i

#data
module m {
typedef 0 t;
typedef long mylong;
};
#errors
2.9;before webidl type;m;typedef
#document
| module m {
| 
| typedef long mylong;
| 
| };

#data
module m {
typedef long 0;
interface x {};
};
#errors
2.14;no webidl identifier;m;typedef
#document
| module m {
| 
| interface x {
| };
| 
| };
